Python: no longer use wrapPythonPrograms
Don't use wrapPythonPrograms anymore.

- If you need a Python environment for development, use `python.devEnv`
or `python.withPackages`. `python.buildEnv` is deprecated in favor of
`python.buildEnv`.
- If you need to use a certain package as an application, you can a) use
the new function `asPythonApplication` to convert a Python "library" to
an application. The function `buildPythonApplication` now uses
`asPythonApplication` as well.

The `asPythonApplication` function creates two additional derivations.
The first is an environment, build with `python.devEnv`. The second is a
minimal derivation symlinking items in /bin of our package but via the
env created with `python.devEnv`.