Python: don't disable support for manylinux1
While Nix doesn't provide the typical libraries at the typical locations
for manylinux1, this is something that is only relevant when using pip
directly. As that's not a supported workflow in Nixpkgs, we may as well
ignore manylinux1 compatibility alltogether. This makes it easier for
those that do want to use pip to install binary wheels, even when there
is a good chance it won't work.