WIP buildPython*: use pythonPath instead of propagatedBuildInputs, fixes #30804
The `buildPython*` functions have been using `propagatedBuildInputs` for
declaring runtime dependencies for a long time. This is now being done
with the previously deprecated `pythonPath`.

Using `propagatedBuildInputs` was done out of convience; by propagating
the build inputs these dependencies would be automatically available
when using the package in another expression. For runtime, the wrapper
generating hook would pick up the propagated build inputs and would
inject them in the entry points and create a wrapper with them as well,
to ensure they are available.

...