buildPython*: don't propagate the interpreter
Python packages propagate Python packages they depend on as well as the
interpreter. The interpreter was added because it comes with a setup
hook that sets `PYTHONPATH`.

This hook can cause trouble when dealing with a combination of Python 2
and 3 packages. The following situations exist:
1) both 2 and 3 libraries. This simply cannot be handled with
`PYTHONPATH`. Use `python.withPackages` instead.
2) both 2 and 3 applications. The fact that Python is used should be
merely an implementation detail. The setup hook should *not* be
executed. Therefore, applications should *not* propagate Python.
3) 2 as a library and 3 as application or the other way around. This is
a common situation. If the application does not propagate the hook *and*
if the application is shielded against `PYTHONPATH`, then we're all
good.

Clearly, it would always be better to *not* propagate the interpreter.
The downside is that, when you use a *library*, so not any of its
executables, you need to add Python as a build input.