WIP: Python: wrap and patch using `requiredPythonModules` instead of `propagatedBuildInputs`
Python programs look up their dependencies dynamically. To support this,
Python executables are wrapped and patched.

To find the dependencies to include when wrapping and patching, the
`wrapPythonPrograms` function would traverse the
`propagatedBuildInputs`. `propagatedBuildInputs` are dependencies that
need to be propagated during *build time*. While that is needed, we do
not want to have *all* propagated build inputs always as run-time
dependencies. Therefore, many derivations instead choose to set
`pythonPath`, which `wrapPythonPrograms` also considers.`pythonPath`,
however, is just an attribute that is not recursed into.

This commit implements `requiredPythonModules`, which was already set in
the `passthru` of Python packages for
`python.buildEnv`/`python.withPackages`. Now, it is also passed to the
builder, and `wrapPythonPrograms` will consider this environment
variable instead of `propagatedBuildInputs`. This change will thus
prevent unwanted bloat in wrappers. It also effectively replaces
`pythonPath`. Note we could have used `pythonPath` for this, but I want
to avoid that name because what we do is not exactly what `PYTHONPATH`
does.

Builders other than `buildPythonPackage` will typically need to pass
something like

requiredPythonModules = with python3.pkgs;
computeRequiredPythonModules [ sphinx ];

for run-time Python dependencies.

This change is a prerequisite to unifying how we deal with wrappers in
Nixpkgs. E.g., now we can look into a common method for dealing with
wrappers for Qt, Python and GApps without getting bloat from Python
packages.

TODO: update documentation