Python interpreters: dynamically create hook
This way we can refer to $out of the interpreter,
and still share the hook easily between interpreters.

Not needed right now, but good as soon as we want to
build up PYTHONPATH differently, by considering also
nix/support/required-python-modules.

In that case, you would only want to process that file if
your interpreter is in that list. This e.g. would
allow us to create versioned (NIX_)PYTHONPATH.