python3Packages.bootstrapped-build: to replace pip
`build` is a lightweight PEP 517 frontend for building wheels that could
replace pip for this purpose.

Until the `installer` project is ready as well, there is not much point
in replacing `pip`.

Why replace `pip` with `build` and `installer`? `pip` has a large
codebase doing many more things, and it vendors its dependencies.